serial: 修正 chip id 的展示方式
chip id 应当作为一个 u8[8] 十六进制展示，而不是作为一个 u64

+++ libcskburn_serial/src/cmd.h
@@ -39,7 +39,7 @@ typedef struct {
 bool cmd_sync(cskburn_serial_device_t *dev, uint16_t timeout);
 
 bool cmd_read_flash_id(cskburn_serial_device_t *dev, uint32_t *id);
-bool cmd_read_chip_id(cskburn_serial_device_t *dev, uint64_t *id);
+bool cmd_read_chip_id(cskburn_serial_device_t *dev, uint8_t *id);
 
 bool cmd_nand_init(cskburn_serial_device_t *dev, uint64_t *size);
